fixed user factory so that it creates a fake model

+++ laravel-app/database/migrations/2023_11_02_170107_create_movies_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('movies', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title');
-            $table->integer('yearReleased');
+            $table->bigInteger('yearReleased');
             $table->decimal('avgRating');
             $table->timestamps();
         });
